Add Alerts Detail Page
