+++ locales/en/plugin__monitoring-plugin.json
@@ -1,22 +1,70 @@
 {
+  "Description": "Description",
+  "Active since": "Active since",
+  "State": "State",
+  "Value": "Value",
+  "Silence alert": "Silence alert",
+  "{{name}} details": "{{name}} details",
+  "Alerts": "Alerts",
+  "Alerting rules": "Alerting rules",
+  "Alerting rule details": "Alerting rule details",
+  "Name": "Name",
+  "Severity": "Severity",
+  "Summary": "Summary",
+  "Message": "Message",
+  "Runbook": "Runbook",
+  "Source": "Source",
+  "For": "For",
+  "Expression": "Expression",
+  "Labels": "Labels",
+  "Active alerts": "Active alerts",
+  "None found": "None found",
+  "View alerting rule": "View alerting rule",
+  "Silences": "Silences",
+  "Silence details": "Silence details",
+  "Matchers": "Matchers",
+  "No matchers": "No matchers",
+  "Last updated at": "Last updated at",
+  "Starts at": "Starts at",
+  "Ends at": "Ends at",
+  "Created by": "Created by",
+  "Comment": "Comment",
+  "Firing alerts": "Firing alerts",
+  "Critical": "Critical",
+  "Warning": "Warning",
+  "Info": "Info",
+  "None": "None",
+  "Alert State": "Alert State",
   "Firing": "Firing",
   "Pending": "Pending",
   "Silenced": "Silenced",
   "Not Firing": "Not Firing",
+  "User": "User",
+  "Platform": "Platform",
+  "Alert state": "Alert state",
+  "Create silence": "Create silence",
+  "Expire {{count}} silence_one": "Expire {{count}} silence",
+  "Expire {{count}} silence_other": "Expire {{count}} silences",
+  "Silence State": "Silence State",
   "Active": "Active",
   "Expired": "Expired",
-  "Ends": "Ends",
-  "Since": "Since",
-  "Critical": "Critical",
-  "Info": "Info",
-  "Warning": "Warning",
-  "None": "None",
+  "Creator": "Creator",
+  "Error loading silences from Alertmanager. Alertmanager may be unavailable.": "Error loading silences from Alertmanager. Alertmanager may be unavailable.",
+  "Error": "Error",
+  "Alerting": "Alerting",
+  "Alert details": "Alert details",
+  "Alerting rule": "Alerting rule",
+  "Silenced by": "Silenced by",
   "Pending: ": "Pending: ",
   "The alert is active but is waiting for the duration that is specified in the alerting rule before it fires.": "The alert is active but is waiting for the duration that is specified in the alerting rule before it fires.",
   "Firing: ": "Firing: ",
   "The alert is firing because the alert condition is true and the optional `for` duration has passed. The alert will continue to fire as long as the condition remains true.": "The alert is firing because the alert condition is true and the optional `for` duration has passed. The alert will continue to fire as long as the condition remains true.",
   "Silenced: ": "Silenced: ",
   "The alert is now silenced for a defined time period. Silences temporarily mute alerts based on a set of label selectors that you define. Notifications will not be sent for alerts that match all the listed values or regular expressions.": "The alert is now silenced for a defined time period. Silences temporarily mute alerts based on a set of label selectors that you define. Notifications will not be sent for alerts that match all the listed values or regular expressions.",
+  "Error loading silences from Alertmanager. Some of the alerts below may actually be silenced.": "Error loading silences from Alertmanager. Some of the alerts below may actually be silenced.",
+  "Ends": "Ends",
+  "Since": "Since",
+  "View in Metrics": "View in Metrics",
   "Critical: ": "Critical: ",
   "The condition that triggered the alert could have a critical impact. The alert requires immediate attention when fired and is typically paged to an individual or to a critical response team.": "The condition that triggered the alert could have a critical impact. The alert requires immediate attention when fired and is typically paged to an individual or to a critical response team.",
   "Warning: ": "Warning: ",
@@ -30,61 +78,13 @@
   "Platform-level alerts relate only to OpenShift namespaces. OpenShift namespaces provide core OpenShift functionality.": "Platform-level alerts relate only to OpenShift namespaces. OpenShift namespaces provide core OpenShift functionality.",
   "User: ": "User: ",
   "User workload alerts relate to user-defined namespaces. These alerts are user-created and are customizable. User workload monitoring can be enabled post-installation to provide observability into your own services.": "User workload alerts relate to user-defined namespaces. These alerts are user-created and are customizable. User workload monitoring can be enabled post-installation to provide observability into your own services.",
-  "View in Metrics": "View in Metrics",
   "Starts": "Starts",
-  "Platform": "Platform",
-  "User": "User",
-  "Name": "Name",
-  "Firing alerts": "Firing alerts",
-  "State": "State",
-  "Creator": "Creator",
-  "Silenced by": "Silenced by",
-  "Alerts": "Alerts",
-  "Alert details": "Alert details",
-  "Silence alert": "Silence alert",
-  "Severity": "Severity",
-  "Description": "Description",
-  "Summary": "Summary",
-  "Message": "Message",
-  "Runbook": "Runbook",
-  "Source": "Source",
-  "Labels": "Labels",
-  "Alerting rule": "Alerting rule",
-  "Active since": "Active since",
-  "Value": "Value",
-  "{{name}} details": "{{name}} details",
-  "Alerting rules": "Alerting rules",
-  "Alerting rule details": "Alerting rule details",
-  "For": "For",
-  "Expression": "Expression",
-  "Active alerts": "Active alerts",
-  "None found": "None found",
+  "Recreate silence": "Recreate silence",
+  "Edit silence": "Edit silence",
   "Expire silence": "Expire silence",
   "Are you sure you want to expire this silence?": "Are you sure you want to expire this silence?",
   "An error occurred": "An error occurred",
   "Cancel": "Cancel",
-  "Recreate silence": "Recreate silence",
-  "Edit silence": "Edit silence",
-  "View alerting rule": "View alerting rule",
-  "Silences": "Silences",
-  "Silence details": "Silence details",
-  "Matchers": "Matchers",
-  "No matchers": "No matchers",
-  "Last updated at": "Last updated at",
-  "Starts at": "Starts at",
-  "Ends at": "Ends at",
-  "Created by": "Created by",
-  "Comment": "Comment",
-  "Error loading silences from Alertmanager. Some of the alerts below may actually be silenced.": "Error loading silences from Alertmanager. Some of the alerts below may actually be silenced.",
-  "Alert State": "Alert State",
-  "Alert state": "Alert state",
-  "Create silence": "Create silence",
-  "Expire {{count}} silence_one": "Expire {{count}} silence",
-  "Expire {{count}} silence_other": "Expire {{count}} silences",
-  "Silence State": "Silence State",
-  "Error loading silences from Alertmanager. Alertmanager may be unavailable.": "Error loading silences from Alertmanager. Alertmanager may be unavailable.",
-  "Error": "Error",
-  "Alerting": "Alerting",
   "Custom time range": "Custom time range",
   "From": "From",
   "To": "To",
